feat: Abstract game creation logic
